Move item event mapping into app-server-protocol
