+++ codex-rs/app-server-protocol/src/protocol/mod.rs
@@ -2,6 +2,7 @@
 // Exposes protocol pieces used by `lib.rs` via `pub use protocol::common::*;`.
 
 pub mod common;
+pub mod event_mapping;
 pub mod item_builders;
 mod mappers;
 mod serde_helpers;
