+++ codex-rs/core-api/Cargo.toml
@@ -13,6 +13,7 @@ path = "src/lib.rs"
 workspace = true
 
 [dependencies]
+codex-app-server-protocol = { workspace = true }
 codex-arg0 = { workspace = true }
 codex-analytics = { workspace = true }
 codex-config = { workspace = true }
